refactor: Update Package component to include rollup-plugin-visualizer
This commit adds the rollup-plugin-visualizer package to the project dependencies in the package.json file. It also updates the vite.config.ts file to include the visualizer plugin in the Vite configuration. This enhancement improves the build process by providing a visual representation of the bundle size and dependencies, helping developers optimize the application's performance and reduce the bundle size.